requirement.txt
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
paho-mqtt==1.5.1
pandas==1.2.5
Pillow==8.2.0
plotly==5.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tenacity==7.0.0
urllib3==1.26.6
Werkzeug==2.0.1
